requirements.txt
adal==1.2.7
alabaster==0.7.12
anyio==3.4.0
applicationinsights==0.11.10
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.9.2
attrs==21.4.0
azure-common==1.1.27
azure-core==1.20.1
azure-graphrbac==0.61.1
azure-identity==1.7.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==19.0.0
azure-mgmt-storage==19.0.0
azureml-automl-core==1.37.0
azureml-core==1.37.0.post1
azureml-dataprep==2.25.2
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.1.1
azureml-dataset-runtime==1.37.0
azureml-pipeline==1.37.0
azureml-pipeline-core==1.37.0
azureml-pipeline-steps==1.37.0
azureml-telemetry==1.37.0
azureml-train-automl-client==1.37.0
azureml-train-core==1.37.0
azureml-train-restclients-hyperdrive==1.37.0
Babel==2.9.1
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
contextlib2==21.6.0
cryptography==3.4.8
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distro==1.6.0
docker==5.0.3
docutils==0.16
dotnetcore2==2.1.22
entrypoints==0.3
fonttools==4.28.5
gitdb==4.0.9
GitPython==3.1.25
greenlet==1.1.2
humanfriendly==9.2
idna==3.3
imagesize==1.3.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
isort==5.10.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-book==0.12.1
jupyter-cache==0.4.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyter-server-mathjax==0.2.3
jupyter-sphinx==0.3.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
jupytext==1.11.5
kiwisolver==1.3.2
knack==0.8.2
latexcodec==2.0.1
lazy-object-proxy==1.7.1
linkify-it-py==1.0.3
markdown-it-py==1.1.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mdit-py-plugins==0.2.8
mistune==0.8.4
msal==1.16.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
myst-nb==0.13.1
myst-parser==0.15.2
nbclient==0.5.9
nbconvert==6.4.0
nbdime==3.1.1
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio==1.5.4
notebook==6.4.6
numpy==1.21.5
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
paramiko==2.9.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pkginfo==1.8.2
platformdirs==2.4.1
portalocker==2.3.2
prometheus-client==0.12.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pyarrow==3.0.0
pyasn1==0.4.8
pybtex==0.24.0
pybtex-docutils==1.0.1
pycparser==2.21
pydata-sphinx-theme==0.7.2
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.12.2
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.0
scikit-learn==1.0.2
scipy==1.7.3
SecretStorage==3.3.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.3.2
sphinx-book-theme==0.1.9
sphinx-comments==0.0.3
sphinx-copybutton==0.4.0
sphinx-external-toc==0.2.3
sphinx-jupyterbook-latex==0.4.6
sphinx-multitoc-numbering==0.1.3
sphinx-panels==0.6.0
sphinx-thebe==0.0.10
sphinx-togglebutton==0.2.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.2.1
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.29
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typed-ast==1.5.1
typing_extensions==4.0.1
uc-micro-py==1.0.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.7.0
